Optimize Integer == (Symbol or String) as a fastpath
Previously `0 == :foo` was much slower than `0 == -1` because the VM
optimizes only the case where the arguments are both Fixnums.

```
require "benchmark"

Benchmark.bmbm do |b|
  b.report { 10000000.times { 0 == -1 } }
  b.report { 10000000.times { 0 == :foo } }
end
```

Before:

```
       user     system      total        real
   0.291631   0.000000   0.291631 (  0.291660)
   1.636855   0.000000   1.636855 (  1.636883)
```

After:

```
       user     system      total        real
   0.290631   0.000000   0.290631 (  0.290632)
   0.291385   0.000000   0.291385 (  0.291387)
```

I encountered this issue when tweaking a program for programming
contest. It used some magic numbers like -1 for special meaning.
Because it is not a good practice, I replaced it with `:special`, but
unfortunately the code became significantly slower (for my case, three
times slower in total).

Note: Interestingly, Symbol == Integer is already fast enough; there is
another fastpath for Object#== which is used for Symbol == Integer
because there is no Symbol#== defined.