file.c: Add `File.read` method as an alias to `IO.read`
`File.read` is a preferable way than `IO.read` when a subprocess
invocation is not needed. However, the rdoc showed only `IO.read`, which
was confusing to users.
This change adds a method definition of `File.read` as an alias to
`IO.read` to list `File.read` in the document.